requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
cssutils==1.0.2
Django==3.1.4
django-cors-headers==3.6.0
django-inlinecss==0.3.0
djangorestframework==3.12.2
future==0.18.2
Pillow==8.0.1
pynliner==0.8.0
python-decouple==3.3
pytz==2020.4
soupsieve==2.1
sqlparse==0.4.1
